+++ tests/continuous/test_brownian_bridge.py
@@ -5,26 +5,26 @@ import pytest
 from stochastic.continuous import BrownianBridge
 
 
-def test_brownian_bridge_str_repr(t, b):
+def test_brownian_bridge_str_repr(b, t):
     if b is not None:
-        instance = BrownianBridge(t, b)
+        instance = BrownianBridge(b, t)
     else:
         instance = BrownianBridge(t)
     assert isinstance(repr(instance), str)
     assert isinstance(str(instance), str)
 
-def test_brownian_bridge_sample(t, b, n, zero, threshold):
+def test_brownian_bridge_sample(b, t, n, zero, threshold):
     if b is not None:
-        instance = BrownianBridge(t, b)
+        instance = BrownianBridge(b, t)
     else:
         instance = BrownianBridge(t)
     s = instance.sample(n, zero)
     assert len(s) == n + int(zero)
     assert s[-1] == pytest.approx(instance.b, threshold)
 
-def test_brownian_bridge_sample_at(t, b, times, threshold):
+def test_brownian_bridge_sample_at(b, t, times, threshold):
     if b is not None:
-        instance = BrownianBridge(t, b)
+        instance = BrownianBridge(b, t)
     else:
         instance = BrownianBridge(t)
     s = instance.sample_at(times)